feat: implemented `Manager` interface & abstract class
