+++ src/main/java/pls_no_shinobu/videostore/errors/ManagerLimitException.java
@@ -1,0 +1,18 @@
+/*
+  RMIT University Vietnam
+  Course: INTE2512 Object-Oriented Programming
+  Semester: 2023A
+  Assessment: Final Project
+  Author: pls_no_shinobu
+*/
+package pls_no_shinobu.videostore.errors;
+
+public class ManagerLimitException extends Exception {
+    public ManagerLimitException() {
+        super();
+    }
+
+    public ManagerLimitException(String message) {
+        super(message);
+    }
+}
